Validate restart with db not removed
Add new option to scripts "no_rm" to avoid deleting all dbs.
Note that user autogenerated tx does not work accross restart so after restart autogenerated tx are not there.

Test:
src/bin/node_settings_run.sh
src/bin/node_settings_run.sh no_rm
=> Restart from 0 since non raft does not have persistent consensused storage

src/bin/node_settings_local_raft_run.sh
src/bin/node_settings_local_raft_run.sh no_rm
=> Restart from last block number

src/bin/node_settings_local_raft_3_run.sh
src/bin/node_settings_local_raft_3_run.sh no_rm
=> Restart from last block number

+++ src/bin/node_settings_local_raft_run.sh
@@ -15,7 +15,11 @@ echo "//-----------------------------//"
 echo "Delete databases"
 echo "//-----------------------------//"
 echo " "
-rm -rf src/db/db/test.* src/wallet/wallet/test.*
+if [ "$1" != "no_rm" ]
+then
+    echo "delete dbs"
+    rm -rf src/db/db/test.* src/wallet/wallet/test.*
+fi
 echo " "
 echo "//-----------------------------//"
 echo "Running nodes for node_settings_local_raft.toml"
